backend/requirements.txt
flask==1.1.1
gunicorn==20.0.4
flask-mongoengine
opencv-python-headless
pillow
mediapipe
numpy
flask-cors==3.0.7
passlib==1.7.4
flask-restplus
Werkzeug==0.16.1
celery
redis
flask_bcrypt
pyjwt==2.1.0
elasticsearch
configparser==5.0.2